chore: UseCase 코드명 변경 및 개선
